--- data/maxine.txt
@@ -1,1 +1,0 @@
-T / 0 / project
