--- data/test/eventtest.txt
@@ -1,1 +1,0 @@
-E / 0 / discussion with friends /  today /  tomorrow
