--- src/test/java/seedu/maxine/stubs/TaskStub.java
@@ -1,16 +1,0 @@
-package seedu.maxine.stubs;
-
-import seedu.maxine.task.Task;
-
-class TaskStub extends Task {
-    private final String data;
-
-    TaskStub(String data) {
-        this.data = data;
-    }
-
-    @Override
-    public String writeToFile() {
-        return data;
-    }
-}
